--- tests/utils/mocks.ts
@@ -1,3 +1,0 @@
-export const FUEL_MNEMONIC =
-  'demand fashion unaware upgrade upon heart bright august panel kangaroo want gaze';
-export const FUEL_WALLET_PASSWORD = '$123Ran123Dom123!';
